Rename client/serverSentEvents.ts
