+++ packages/lesswrong/client/serverSentEventsClient.ts
