+++ packages/lesswrong/client/start.tsx
@@ -7,7 +7,7 @@ import type { TimeOverride } from '../lib/utils/timeUtil';
 import { createApolloClient } from './apolloClient';
 import { fmCrosspostBaseUrlSetting } from "../lib/instanceSettings";
 import { populateComponentsAppDebug } from '../lib/vulcan-lib';
-import { subscribeToNotifications } from "./serverSentEvents";
+import { subscribeToNotifications } from "./serverSentEventsClient";
 
 onStartup(() => {
   populateComponentsAppDebug();
